backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.31
psycopg[binary]==3.2.1
pydantic-settings==2.3.4
python-jose==3.3.0
requests==2.32.3
openpyxl==3.1.5
